Fix major issue with recycler view layout, other optimizations
Removed all redundant instances / declarations that we used for the singleton instance. We can just declare a private variable and store it when the view loads
Moved some test data sets into the singleton instance to help with future testing so the datasets are freely available
Recycler view issue was due to me using the existing viewholder property instead of binding it to something else to keep track of what the viewholder should actually contain. Moved all of my important items (color, etc) to be stored in a int[] array
Removed unecessary banner.kill calls and due to above, was able to fix banners showing up in the wrong indices
GDPR dialogue is now prettier and doesn't falsly report that consent was not given. I was doing a logical check for it, but still rendered the wrong state because I referenced an older instance of the GV.gdpr status (since I saved the state, this is related to the 1st point) and thus would only render correctly after view is redrawn
Added icons, because that's always nice to have
Upgraded the gradle file to stop using compile to stop the warning